+++ packages/components/src/Pre.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default props => (
+  <pre
+    {...props}
+    style={{
+      fontFamily: 'Menlo, monospace',
+      fontSize: 18,
+    }}
+  />
+)